Update app version, improve context date handling, and adjust I/O digest parameters. (#207)
This commit introduces a version increase from 0.6.1 to 0.6.2 and a significant overhaul of the date handling in context. The handling of 'from_date' and 'to_date' in context across different MixIn classes has been refactored using the new ContextFromToDateMixin. The I/O digest parameters in DigestContextMixIn were adjusted, with process_ratios and process_roles set to True. Unnecessary spacing in the balance_sheet_statement.html template was also cleaned up.